[WIP, DIRTY] added Virtual Keyboard & config.json
This commit is quite dirty, but functional. Added linux/uinput
to the mix, for sending keypresses to a virtual keyboard with
no noticable latency (unlike xdotool). I also added a transpile
script which can build the config.cpp for convenience, however
the config.json is incomplete and auto/install_full.rb needs to
be updated.