[BUGFIX] us.new2 was corrected, previously unparseable.
Had a mistake in the us symbols file (for the symbolic variant)
which caused setxkbmap to fail. Apparently each key needs to be
mapped to at least 3 symbols, and the last element cannot be
succeded by a comma.

Also added error handling and warning message to install_us_keyboard_layout
if it fails.