Better x11 xkb us symbolic keyboard layout
Previously pressing »Caps+d« and similar would send »Escape« to
the active window. Most windows ignore this keystroke for whatever reason,
but Firefox and chrome do not, which causes problems on various websites.
Using »NoSymbol« or »U0000« doesn't work, it will send the base key -in
this case »d«.
Replacing »Escape« with »Caps« however will actually send »Caps«, which
is basically ignored by X11 but not by the browser, it will interpret
the keystroke as »Caps« but of course this has no further meaning,
so it does nothing, which is exactly the desired behaviour.